requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.
django-enum-choices==2.1.3
importlib-metadata==4.8.2
kombu==5.2.2
prompt-toolkit==3.0.22
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
psycopg2==2.9.2
ebaysdk==2.2.0
django-rest-swagger==2.2.0
